requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dj-database-url==1.0.0
Django==4.1.1
django-heroku==0.3.1
djangorestframework==3.14.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.19.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.1